pyproject.toml
[project]
name = "rag-test"
version = "0.1.0"
description = "Тестовая RAG система"
requires-python = ">=3.10"

dependencies = [
    "sentence-transformers",
    "faiss-cpu",
    "langchain == 0.3.27",
    "pypdf",
    "numpy",
    'docx2txt'
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"